backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
reportlab
pytesseract
pillow
pdf2image
vosk
soundfile
